% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_frequency}
\alias{allele_frequency}
\title{Allele Frequency}
\format{
A dataframe with 25,000 rows and 4 columns
}
\source{
\url{https://www.frontiersin.org/journals/genetics/articles/10.3389/fgene.2025.1535368/full#h5}
}
\usage{
data("example2")
}
\description{
A DataFrame is used for the parameter data in ccGenomicPoints(). Example can be found in "4.2 Example 2: comparison of two pieces of code that use circlizePlus and circlize to implement the same requirements, respectively"
}
\keyword{datasets}
