% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCell-class}
\alias{ccCell-class}
\title{S4 class ccCell}
\description{
S4 class ccCell
}
\section{Slots}{

\describe{
\item{\code{sector.index}}{character.}

\item{\code{geoms}}{list.}
}}

\examples{
NULL
}
