% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCellGeom-class}
\alias{ccCellGeom-class}
\title{S4 class ccCellGeom}
\description{
S4 class ccCellGeom
}
\section{Slots}{

\describe{
\item{\code{func}}{character.}

\item{\code{params}}{list.}
}}

\examples{
NULL
}
