% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCells-class}
\alias{ccCells-class}
\title{S4 class ccCells}
\description{
S4 class ccCells
}
\examples{
library(circlizePlus)
n = 1000
df = data.frame(sectors = sample(letters[1:8], n, replace = TRUE),
                x = rnorm(n), y = runif(n))
par1=ccPar("track.height" = 0.1)
cc=ccPlot(sectors = df$sectors, x = df$x) + par1
track1 = ccTrack(sectors = df$sectors, y = df$y,
                 panel.fun = function(x, y) {
                   circos.text(CELL_META$xcenter,
                               CELL_META$cell.ylim[2] + mm_y(5),
                               CELL_META$sector.index)
                   circos.axis(labels.cex = 0.6)
                 })
cell1 = ccCells(sector.index = 'a')
track1 = track1 + cell1
cc = cc + track1
cc
}
