% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ccGenomicLabels}
\alias{ccGenomicLabels}
\title{Object generator for S4 class ccGenomicTrack}
\usage{
ccGenomicLabels(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicLabels]{circlize::circos.genomicLabels}}
  \describe{
    \item{\code{bed}}{A data frame in bed format.
}
    \item{\code{labels}}{A vector of labels corresponding to rows in \code{bed}.
}
    \item{\code{labels.column}}{If the label column is already in \code{bed}, the index for this column in \code{bed}.
}
    \item{\code{facing}}{fFacing of the labels. The value can only be \code{"clockwise"} or \code{"reverse.clockwise"}.
}
    \item{\code{niceFacing}}{Whether automatically adjust the facing of the labels.
}
    \item{\code{col}}{Color for the labels.
}
    \item{\code{cex}}{Size of the labels.
}
    \item{\code{font}}{Font of the labels.
}
    \item{\code{padding}}{Padding of the labels, the value is the ratio to the height of the label.
}
    \item{\code{connection_height}}{Height of the connection track.
}
    \item{\code{line_col}}{Color for the connection lines.
}
    \item{\code{line_lwd}}{Line width for the connection lines.
}
    \item{\code{line_lty}}{Line type for the connectioin lines.
}
    \item{\code{labels_height}}{Height of the labels track.
}
    \item{\code{side}}{Side of the labels track, is it in the inside of the track where the regions are marked?
}
    \item{\code{labels.side}}{Same as \code{side}. It will replace \code{side} in the future versions.
}
    \item{\code{track.margin}}{Bottom and top margins.
}
  }}
}
\value{
Object \linkS4class{ccGenomicTrack}
}
\description{
Object \linkS4class{ccGenomicTrack} will call the function \link{circos.genomicLabels} while drawing.
}
\examples{
library(circlizePlus)
bed = generateRandomBed(nr = 50, fun = function(k) sample(letters, k, replace = TRUE))
bed[1, 4] = "aaaaa"
cc = ccPlot(initMode = "initializeWithIdeogram", plotType = NULL)
t1 = ccGenomicLabels(bed, labels.column = 4, side = "outside",
                     col = as.numeric(factor(bed[[1]])), line_col = as.numeric(factor(bed[[1]])))
cc + t1
circos.clear()
}
