% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\docType{class}
\name{ccGenomicLink-class}
\alias{ccGenomicLink-class}
\title{S4 class ccGenomicLink}
\description{
S4 class ccGenomicLink
}
\section{Slots}{

\describe{
\item{\code{func}}{character. Normally it is "circos.genomicLink".}

\item{\code{params}}{list. A \strong{named} list that stores the parameters of the function \link[circlize:circos.genomicLink]{circlize::circos.genomicLink} called by the backend.}
}}

\examples{
NULL
}
