% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{class}
\name{ccGenomicTrack-class}
\alias{ccGenomicTrack-class}
\title{S4 class ccGenomicTrack}
\description{
S4 class ccGenomicTrack
}
\section{Slots}{

\describe{
\item{\code{func}}{character.}

\item{\code{params}}{list.}

\item{\code{trackGeoms}}{list.}

\item{\code{cells}}{list.}
}}

\examples{
library(circlizePlus)
cc = ccPlot(initMode = "initializeWithIdeogram",chromosome.index = "chr1", plotType = NULL)
bed = generateRandomBed(nr = 300)
t1 = ccGenomicTrack(bed, panel.fun = function(region, value, ...) {
  circos.genomicPoints(region, value, pch = 16, cex = 0.5, ...)
})
cc+t1
circos.clear()
}
