% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\docType{class}
\name{ccHeatmapLink-class}
\alias{ccHeatmapLink-class}
\title{S4 class ccHeatmapLink}
\description{
S4 class ccHeatmapLink
}
\section{Slots}{

\describe{
\item{\code{func}}{character. Normally it is "circos.heatmap.link".}

\item{\code{params}}{list. A \strong{named} list that stores the parameters of the function \link[circlize:circos.heatmap.link]{circlize::circos.heatmap.link} called by the backend.}
}}

\examples{
NULL
}
