% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\docType{class}
\name{ccLink-class}
\alias{ccLink-class}
\title{S4 class ccLink}
\description{
S4 class ccLink
}
\section{Slots}{

\describe{
\item{\code{func}}{character.}

\item{\code{params}}{list.}
}}

\examples{
NULL
}
