% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{class}
\name{ccPar-class}
\alias{ccPar-class}
\title{Object generator for S4 class ccPar}
\description{
Object generator for S4 class ccPar
}
\section{Slots}{

\describe{
\item{\code{params}}{params list. A \strong{named} list that stores the parameters of the function \link[circlize:circos.par]{circlize::circos.par} called by the backend.}
}}

\examples{
NULL
}
