% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{ccPar}
\alias{ccPar}
\title{Object generator for S4 class ccPar}
\usage{
ccPar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.par]{circlize::circos.par}}
  \describe{
    \item{\code{RESET}}{reset to default values
}
    \item{\code{READ.ONLY}}{please ignore
}
    \item{\code{LOCAL}}{please ignore
}
    \item{\code{ADD}}{please ignore
}
  }}
}
\value{
Object \linkS4class{ccPar}
}
\description{
Object \linkS4class{ccPar} will call the function \link{circos.par} while drawing.
}
\examples{
NULL
}
