% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccPoints}
\alias{ccPoints}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccPoints(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.points]{circlize::circos.points}}
  \describe{
    \item{\code{x}}{Data points on x-axis, measured in "current" data coordinate}
    \item{\code{y}}{Data points on y-axis, measured in "current" data coordinate}
    \item{\code{sector.index}}{Index for the sector}
    \item{\code{track.index}}{Index for the track}
    \item{\code{pch}}{Point type}
    \item{\code{col}}{Point color}
    \item{\code{cex}}{Point size}
    \item{\code{bg}}{backgrond of points}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.points} while drawing.
}
\examples{
NULL
}
