% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccRect}
\alias{ccRect}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccRect(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.rect]{circlize::circos.rect}}
  \describe{
    \item{\code{xleft}}{x for the left bottom points}
    \item{\code{ybottom}}{y for the left bottom points}
    \item{\code{xright}}{x for the right top points}
    \item{\code{ytop}}{y for the right top points}
    \item{\code{sector.index}}{Index for the sector}
    \item{\code{track.index}}{Index for the track}
    \item{\code{rot}}{Rotation of the rectangles. The value is measured clockwise in degree. Rotation is relative to the center of the rectangles.}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.rect} while drawing.
}
\examples{
NULL
}
