% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{class}
\name{ccTrack-class}
\alias{ccTrack-class}
\title{S4 class ccTrack}
\description{
S4 class ccTrack
}
\section{Slots}{

\describe{
\item{\code{func}}{character.}

\item{\code{params}}{list.}

\item{\code{trackGeoms}}{list.}

\item{\code{cells}}{list.}
}}

\examples{
NULL
}
