% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ccTrack}
\alias{ccTrack}
\title{Object generator for S4 class ccTrack}
\usage{
ccTrack(
  sectors = NULL,
  x = NULL,
  y = NULL,
  ylim = NULL,
  force.ylim = NULL,
  track.index = NULL,
  track.height = NULL,
  track.margin = NULL,
  cell.padding = NULL,
  bg.col = NULL,
  bg.border = NULL,
  bg.lty = NULL,
  bg.lwd = NULL,
  panel.fun = NULL,
  factors = NULL
)
}
\arguments{
\item{sectors}{A \code{\link{factor}} or a character vector which represents categories of data, if it is \code{NULL}, then it uses all sector index.}

\item{x}{Data on x-axis. It is only used if \code{panel.fun} is set.}

\item{y}{Data on y-axis}

\item{ylim}{Range of data on y-axis}

\item{force.ylim}{Whether to force all cells in the track to share the same \code{ylim}. Normally, all cells on a same track should have same \code{ylim}.}

\item{track.index}{Index for the track which is going to be created/updated. If the specified track has already been created, this function just updated corresponding track with new plot. If the specified track is \code{NULL} or has not been created, this function just creates it. Note the value for this argument should not exceed maximum track index plus 1.}

\item{track.height}{Height of the track. It is the percentage to the radius of the unit circles. The value can be set by \code{\link[circlize]{uh}} to an absolute unit. If updating a track (with proper \code{track.index} value), this argument is ignored.}

\item{track.margin}{only affect current track}

\item{cell.padding}{only affect current track}

\item{bg.col}{Background color for the plotting regions. It can be vector which has the same length of sectors.}

\item{bg.border}{Color for the border of the plotting regions. It can be vector which has the same length of sectors.}

\item{bg.lty}{Line style for the border of the plotting regions. It can be vector which has the same length of sectors.}

\item{bg.lwd}{Line width for the border of the plotting regions. It can be vector which has the same length of sectors.}

\item{panel.fun}{Panel function to add graphics in each cell, see "details" section and vignette for explanation.}

\item{factors}{The same as \code{sectors}. It will be removed in future versions. }
}
\value{
Object \linkS4class{ccTrack}
}
\description{
Object \linkS4class{ccTrack} will call the function \link{circos.trackPlotRegion} while drawing.
}
\examples{
NULL
}
