% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\docType{class}
\name{ccTrackGeom-class}
\alias{ccTrackGeom-class}
\title{S4 class ccTrackGeom}
\description{
S4 class ccTrackGeom
}
\section{Slots}{

\describe{
\item{\code{func}}{character.}

\item{\code{params}}{list.}
}}

\examples{
NULL
}
