% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ccTrackHist}
\alias{ccTrackHist}
\title{Object generator for S4 class ccTrack}
\usage{
ccTrackHist(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.trackHist]{circlize::circos.trackHist}}
  \describe{
    \item{\code{sectors}}{A \code{\link{factor}} or a character vector which represents the categories of data}
    \item{\code{factors}}{The same as \code{sectors}. It will be removed in future versions. }
    \item{\code{x}}{Data on the x-axis}
    \item{\code{track.index}}{Index for the track which is going to be updated. Setting it to \code{NULL} means creating the plotting regions in the next newest track.}
    \item{\code{track.height}}{Height of the track. It is the percentage to the radius of the unit circle. If to update a track, this argument is disabled.}
    \item{\code{ylim}}{Ranges on y-direction. By default, \code{ylim} is calculated automatically.}
    \item{\code{force.ylim}}{Whether to force all cells in the track to share the same \code{ylim}.}
    \item{\code{col}}{Filled color for histogram}
    \item{\code{border}}{Border color for histogram}
    \item{\code{lty}}{Line style for histogram}
    \item{\code{lwd}}{Line width for histogram}
    \item{\code{bg.col}}{Background color for the plotting regions}
    \item{\code{bg.border}}{Color for the border of the plotting regions}
    \item{\code{bg.lty}}{Line style for the border of the plotting regions}
    \item{\code{bg.lwd}}{Line width for the border of the plotting regions}
    \item{\code{breaks}}{see \code{\link[graphics]{hist}}}
    \item{\code{include.lowest}}{see \code{\link[graphics]{hist}}}
    \item{\code{right}}{see \code{\link[graphics]{hist}}}
    \item{\code{draw.density}}{whether draw density lines instead of histogram bars.}
    \item{\code{area}}{whether to fill the area below the density lines. If it is set to \code{TRUE}, \code{col} controls the filled color in the area and \code{border} controls color of the line.}
    \item{\code{bin.size}}{size of the bins of the histogram}
  }}
}
\value{
Object \linkS4class{ccTrack}
}
\description{
Object \linkS4class{ccTrack} will call the function \link{circos.trackHist} while drawing.
}
\examples{
n = 1000
df = data.frame(sectors = sample(letters[1:8], n, replace = TRUE),
                x = rnorm(n), y = runif(n))
library(circlizePlus)
par1=ccPar("track.height" = 0.1)
cc=ccPlot(sectors = df$sectors, x = df$x) + par1;bgcol = rep(c("#EFEFEF", "#CCCCCC"), 4)
track2 = ccTrackHist(df$sectors, df$x, bin.size = 0.2, bg.col = bgcol, col = NA)
cc=cc+track2
cc
circos.clear()
}
