% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\name{ccTrackPoints}
\alias{ccTrackPoints}
\title{Add points on all sections of a single track.}
\usage{
ccTrackPoints(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.trackPoints]{circlize::circos.trackPoints}}
  \describe{
    \item{\code{sectors}}{A \code{\link{factor}} or a character vector which represents the categories of data}
    \item{\code{factors}}{The same as \code{sectors}. It will be removed in future versions. }
    \item{\code{x}}{Data points on x-axis}
    \item{\code{y}}{Data points on y-axis}
    \item{\code{track.index}}{Index for the track}
    \item{\code{pch}}{Point type}
    \item{\code{col}}{Point color}
    \item{\code{cex}}{Point size}
    \item{\code{bg}}{backgrond color}
  }}
}
\value{
Object \linkS4class{ccTrackGeom}
}
\description{
Object \linkS4class{ccTrackGeom} will call the function \link[circlize:circos.trackPoints]{circlize::circos.trackPoints} while drawing.
}
\examples{
NULL
}
