% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\name{ccTrackText}
\alias{ccTrackText}
\title{Object generator for S4 class ccTrackGeom}
\usage{
ccTrackText(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.trackText]{circlize::circos.trackText}}
  \describe{
    \item{\code{sectors}}{A \code{\link{factor}} or a character vector which represents the categories of data}
    \item{\code{factors}}{The same as \code{sectors}. It will be removed in future versions. }
    \item{\code{x}}{Data points on x-axis}
    \item{\code{y}}{Data points on y-axis}
    \item{\code{labels}}{Labels}
    \item{\code{track.index}}{Index for the track}
    \item{\code{direction}}{deprecated, use \code{facing} instead.}
    \item{\code{facing}}{Facing of text}
    \item{\code{niceFacing}}{Should the facing of text be adjusted to fit human eyes?}
    \item{\code{adj}}{Adjustment for text}
    \item{\code{cex}}{Font size}
    \item{\code{col}}{Font color}
    \item{\code{font}}{Font style}
  }}
}
\value{
Object \linkS4class{ccTrackGeom}
}
\description{
Object \linkS4class{ccTrackGeom} will call the function \link{circos.trackText} while drawing.
}
\examples{
NULL
}
