% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-set}
\alias{data-set}
\alias{allele_frequency}
\alias{copy_number}
\alias{junctions}
\alias{r1}
\alias{r2}
\title{Data set in circlizePlus}
\format{
An object of class \code{data.frame} with 25000 rows and 4 columns.

An object of class \code{data.frame} with 24788 rows and 5 columns.

An object of class \code{data.frame} with 2414 rows and 28 columns.

An object of class \code{data.frame} with 2414 rows and 3 columns.

An object of class \code{data.frame} with 2414 rows and 3 columns.
}
\source{
\url{https://www.frontiersin.org/journals/genetics/articles/10.3389/fgene.2025.1535368/full#h5}
}
\usage{
data(example2)

data(example2)

data(example2)

data(example2)

data(example2)
}
\description{
Example can be found in "4.2 Example 2: comparison of two pieces of code that use circlizePlus and circlize to implement the same requirements, respectively"
\if{html}{\out{<br>}}\if{html}{\out{<br>}}allele_frequency: A dataframe with 25,000 rows and 4 columns
\if{html}{\out{<br>}}copy_number: A dataframe with 24,788 rows and 5 columns
\if{html}{\out{<br>}}junctions: A dataframe with 2414 rows and 28 columns
\if{html}{\out{<br>}}r1: A dataframe with 2414 rows and 3 columns
\if{html}{\out{<br>}}r2: A dataframe with 2414 rows and 3 columns
}
\examples{
\dontrun{
data(example2)
allele_frequency
copy_number
junctions
r1
r2
}
}
\keyword{datasets}
