% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r1}
\alias{r1}
\title{R1}
\format{
A dataframe with 2414 rows and 3 columns
}
\source{
\url{https://www.frontiersin.org/journals/genetics/articles/10.3389/fgene.2025.1535368/full#h5}
}
\usage{
data("example2")
}
\description{
A DataFrame is used for the parameter data in ccGenomicLink(). Example can be found in "4.2 Example 2: comparison of two pieces of code that use circlizePlus and circlize to implement the same requirements, respectively"
}
\keyword{datasets}
