% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addition-rules}
\alias{addition-rules}
\alias{+,ccPlot,ccPar-method}
\alias{+,ccPlot,ccTrack-method}
\alias{+,ccPlot,ccLink-method}
\alias{+,ccTrack,ccTrackGeom-method}
\alias{+,ccTrack,ccCells-method}
\alias{+,ccTrack,ccCell-method}
\alias{+,ccCell,ccCellGeom-method}
\alias{+,ccCells,ccCellGeom-method}
\alias{+,ccHeatmap,ccPar-method}
\alias{+,ccHeatmap,ccTrack-method}
\alias{+,ccHeatmap,ccLink-method}
\title{Addition rules in circlizePlus}
\usage{
\S4method{+}{ccPlot,ccPar}(e1, e2)

\S4method{+}{ccPlot,ccTrack}(e1, e2)

\S4method{+}{ccPlot,ccLink}(e1, e2)

\S4method{+}{ccTrack,ccTrackGeom}(e1, e2)

\S4method{+}{ccTrack,ccCells}(e1, e2)

\S4method{+}{ccTrack,ccCell}(e1, e2)

\S4method{+}{ccCell,ccCellGeom}(e1, e2)

\S4method{+}{ccCells,ccCellGeom}(e1, e2)

\S4method{+}{ccHeatmap,ccPar}(e1, e2)

\S4method{+}{ccHeatmap,ccTrack}(e1, e2)

\S4method{+}{ccHeatmap,ccLink}(e1, e2)
}
\arguments{
\item{e1}{A object defined in circlizePlus}

\item{e2}{A object defined in circlizePlus}
}
\value{
A object defined in circlizePlus
}
\description{
ccPlot(contain n ccTracks)+ccTrack=ccPlot(contain n+1 ccTracks),n>=0 \if{html}{\out{<br>}}
ccPlot(contain n ccLinks)+ccLink=ccPlot(contain n+1 ccLinks),n>=0 \if{html}{\out{<br>}}
ccTrak(contain n ccTrakGeoms)+ccTrackGeom=ccTrack(contain n+1 ccTrackGeoms),n>=0 \if{html}{\out{<br>}}
ccTrack(contain n ccCells)+ccCell=ccTrack(contain n+1 ccCells),n>=0 \if{html}{\out{<br>}}
ccCell(contain n ccCellGeoms)+ccCellGeom=ccCell(contain n+1 ccCellGeoms),n>=0 \if{html}{\out{<br>}}
}
\examples{
NULL

}
