% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccBoxplot}
\alias{ccBoxplot}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccBoxplot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.boxplot]{circlize::circos.boxplot}}
  \describe{
    \item{\code{value}}{A numeric vector, a matrix or a list. If it is a matrix, boxplots are made by columns (each column is a box).}
    \item{\code{pos}}{Positions of the boxes.}
    \item{\code{outline}}{Whether to draw outliers.}
    \item{\code{box_width}}{Width of boxes.}
    \item{\code{col}}{Filled color of boxes.}
    \item{\code{border}}{Color for the border as well as the quantile lines.}
    \item{\code{lwd}}{Line width.}
    \item{\code{lty}}{Line style}
    \item{\code{cex}}{Point size.}
    \item{\code{pch}}{Point type.}
    \item{\code{pt.col}}{Point color.}
    \item{\code{sector.index}}{Index of sector.}
    \item{\code{track.index}}{Index of track.}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.boxplot} while drawing.
}
\examples{
NULL
}
