% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCell-class}
\alias{ccCell-class}
\title{S4 class ccCell}
\description{
A cell container that belongs to a particular sector.
}
\section{Slots}{

\describe{
\item{\code{sector.index}}{character. It is the index that corresponds to the sector.}

\item{\code{geoms}}{list. The elements in the list should all be of type \linkS4class{ccCellGeom} or \linkS4class{ccGenomicCellGeom}.}
}}

\examples{
NULL
}
