% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCellGeom-class}
\alias{ccCellGeom-class}
\title{S4 class ccCellGeom}
\description{
Objectified representation of the R package circlize's plotting functions and corresponding parameters at the cell level.
}
\section{Slots}{

\describe{
\item{\code{func}}{character. The name of the plot function in the R package circlize.}

\item{\code{params}}{list. When the function corresponding to the parameter param is called, it represents the argument of this function.}
}}

\examples{
NULL
}
