% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\docType{class}
\name{ccCells-class}
\alias{ccCells-class}
\title{S4 class ccCells}
\description{
A list of multiple ccCell. Any ccCellGeom and ccCells are added together as if they were added to each ccCell contained in the ccCells.
}
\examples{
NULL
}
