% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccDendrogram}
\alias{ccDendrogram}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccDendrogram(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.dendrogram]{circlize::circos.dendrogram}}
  \describe{
    \item{\code{dend}}{A \code{\link[stats]{dendrogram}} object.}
    \item{\code{facing}}{Is the dendromgrams facing inside to the circle or outside?}
    \item{\code{max_height}}{Maximum height of the dendrogram. This is important if more than one dendrograms are drawn in one track and making them comparable. The height of a dendrogram can be obtained by \code{attr(dend, "height")}.}
    \item{\code{use_x_attr}}{Whether use the \code{x} attribute to determine node positions in the dendrogram, used internally.}
    \item{\code{sector.index}}{Index of sector.}
    \item{\code{track.index}}{Index of track.}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.dendrogram} while drawing.
}
\examples{
load(system.file(package = "circlize", "extdata", "bird.orders.RData"))
dend = as.dendrogram(hc)
dend = color_branches(dend, k = 6, col = 1:6)
require(dendextend)
library(circlizePlus)
cc=ccPlot(sectors=letters[1:4], xlim = c(0, 10))
track=ccTrack(ylim = c(0, 1))
cell=ccCell(sector.index = 'a')+ccDendrogram(dend=dend, max_height = attr(dend, "height"))
track=track+cell
cc+track
}
