% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccGenomicAxis}
\alias{ccGenomicAxis}
\title{Object generator for S4 class ccGenomicCellGeom}
\usage{
ccGenomicAxis(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicAxis]{circlize::circos.genomicAxis}}
  \describe{
    \item{\code{h}}{Position of the axes. "top" or "bottom".
}
    \item{\code{major.at}}{Major breaks. If \code{major.at} is set, \code{major.by} is ignored.
}
    \item{\code{labels}}{labels corresponding to \code{major.at}. If \code{labels} is set, \code{major.at} must be set.
}
    \item{\code{major.by}}{Increment of major ticks. It is calculated automatically if the value is not set (about every 10 degrees there is a major tick).
}
    \item{\code{tickLabelsStartFromZero}}{Whether axis tick labels start from 0? This will only affect the axis labels while not affect x-values in cells.
}
    \item{\code{labels.cex}}{The font size for the axis tick labels.
}
    \item{\code{sector.index}}{Index for the sector
}
    \item{\code{track.index}}{Index for the track
}
  }}
}
\value{
Object \linkS4class{ccGenomicCellGeom}
}
\description{
Object \linkS4class{ccGenomicCellGeom} will call the function \link{circos.genomicAxis} while drawing.
}
\examples{
NULL
}
