% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ccGenomicIdeogram}
\alias{ccGenomicIdeogram}
\title{Object generator for S4 class ccGenomicTrack}
\usage{
ccGenomicIdeogram(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicIdeogram]{circlize::circos.genomicIdeogram}}
  \describe{
    \item{\code{cytoband}}{A data frame or a file path, pass to \code{\link[circlize]{read.cytoband}}.
}
    \item{\code{species}}{Abbreviations of the genome, pass to \code{\link[circlize]{read.cytoband}}.
}
    \item{\code{track.height}}{Height of the ideogram track.
}
    \item{\code{track.margin}}{Margins for the track.
}
  }}
}
\value{
Object \linkS4class{ccGenomicTrack}
}
\description{
Object \linkS4class{ccGenomicTrack} will call the function \link{circos.genomicIdeogram} while drawing.
}
\examples{
library(circlizePlus)
cc = ccPlot(initMode = "initializeWithIdeogram",chromosome.index = "chr1", plotType = NULL)
human_cytoband = read.cytoband(species = "hg19")$df
t2=ccGenomicIdeogram(human_cytoband)
cc+t2
circos.clear()
}
