% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccGenomicLines}
\alias{ccGenomicLines}
\title{Object generator for S4 class ccGenomicCellGeom}
\usage{
ccGenomicLines(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicLines]{circlize::circos.genomicLines}}
  \describe{
    \item{\code{region}}{A data frame contains 2 column which correspond to start positions and end positions.
}
    \item{\code{value}}{A data frame contains values and other information.
}
    \item{\code{numeric.column}}{Which column in \code{value} data frame should be taken as y-value.
 If it is not defined, the whole numeric columns in \code{value} will be taken.
}
    \item{\code{sector.index}}{Index of sector.
}
    \item{\code{track.index}}{Index of track.
}
    \item{\code{posTransform}}{Self-defined function to transform genomic positions, see \code{\link[circlize]{posTransform.default}} for explaination.
}
    \item{\code{col}}{col of lines/areas. If there are more than one numeric column, the length of \code{col} can be either one or number of numeric columns.
 If there is only one numeric column and type is either \code{segment} or \code{h}, 
 the length of \code{col} can be either one or number of rows of \code{region}.
 pass to \code{\link[circlize]{circos.lines}}
}
    \item{\code{lwd}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{lty}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{type}}{There is an additional option \code{segment} which plot segment lines from start position to end position. Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}. 
}
    \item{\code{area}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{area.baseline}}{Deprecated, use \code{baseline} instead.
}
    \item{\code{baseline}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{border}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{pt.col}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{cex}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
    \item{\code{pch}}{Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.lines}}.
}
  }}
}
\value{
Object \linkS4class{ccGenomicCellGeom}
}
\description{
Object \linkS4class{ccGenomicCellGeom} will call the function \link{circos.genomicLines} while drawing.
}
\examples{
NULL
}
