% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{ccGenomicLink}
\alias{ccGenomicLink}
\title{Object generator for S4 class ccGenomicLink}
\usage{
ccGenomicLink(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicLink]{circlize::circos.genomicLink}}
  \describe{
    \item{\code{region1}}{A data frame in bed format.
}
    \item{\code{region2}}{A data frame in bed format.
}
    \item{\code{rou}}{Pass to \code{\link[circlize]{circos.link}}.
}
    \item{\code{rou1}}{Pass to \code{\link[circlize]{circos.link}}.
}
    \item{\code{rou2}}{Pass to \code{\link[circlize]{circos.link}}.
}
    \item{\code{col}}{Pass to \code{\link[circlize]{circos.link}}, length can be either one or nrow of \code{region1}.
}
    \item{\code{lwd}}{Pass to \code{\link[circlize]{circos.link}}, length can be either one or nrow of \code{region1}.
}
    \item{\code{lty}}{Pass to \code{\link[circlize]{circos.link}}, length can be either one or nrow of \code{region1}.
}
    \item{\code{border}}{Pass to \code{\link[circlize]{circos.link}}, length can be either one or nrow of \code{region1}.
}
  }}
}
\value{
Object \linkS4class{ccGenomicLink}
}
\description{
Object \linkS4class{ccGenomicLink} will call the function \link{circos.genomicLink} while drawing.
}
\examples{
NULL
}
