% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccGenomicPoints}
\alias{ccGenomicPoints}
\title{Object generator for S4 class ccGenomicCellGeom}
\usage{
ccGenomicPoints(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicPoints]{circlize::circos.genomicPoints}}
  \describe{
    \item{\code{region}}{A data frame contains 2 columns which correspond to start positions and end positions.
}
    \item{\code{value}}{A data frame contains values and other information.
}
    \item{\code{numeric.column}}{Which column in \code{value} data frame should be taken as y-value.
 If it is not defined, the whole numeric columns in \code{value} will be taken.
}
    \item{\code{sector.index}}{Index of sector.
}
    \item{\code{track.index}}{Index of track.
}
    \item{\code{posTransform}}{Self-defined function to transform genomic positions, see \code{\link[circlize]{posTransform.default}} for explanation
}
    \item{\code{col}}{Color of points. If there is only one numeric column, the length of \code{col} can be either one or number of rows of \code{region}.
 If there are more than one numeric column, the length of \code{col} can be either one or number of numeric columns.
 Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{pch}}{Type of points. Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{cex}}{Size of points. Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{bg}}{Background colors for points.
}
  }}
}
\value{
Object \linkS4class{ccGenomicCellGeom}
}
\description{
Object \linkS4class{ccGenomicCellGeom} will call the function \link{circos.genomicPoints} while drawing.
}
\examples{
load(system.file(package = "circlize", "extdata", "bird.orders.RData"))
dend = as.dendrogram(hc)
dend = color_branches(dend, k = 6, col = 1:6)
require(dendextend)
library(circlizePlus)
cc=ccPlot(sectors=letters[1:4], xlim = c(0, 10))
track=ccTrack(ylim = c(0, 1))
cell=ccCell(sector.index = 'a')+ccDendrogram(dend=dend, max_height = attr(dend, "height"))
track=track+cell
cc+track
}
