% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccGenomicPoints}
\alias{ccGenomicPoints}
\title{Object generator for S4 class ccGenomicCellGeom}
\usage{
ccGenomicPoints(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicPoints]{circlize::circos.genomicPoints}}
  \describe{
    \item{\code{region}}{A data frame contains 2 columns which correspond to start positions and end positions.
}
    \item{\code{value}}{A data frame contains values and other information.
}
    \item{\code{numeric.column}}{Which column in \code{value} data frame should be taken as y-value.
 If it is not defined, the whole numeric columns in \code{value} will be taken.
}
    \item{\code{sector.index}}{Index of sector.
}
    \item{\code{track.index}}{Index of track.
}
    \item{\code{posTransform}}{Self-defined function to transform genomic positions, see \code{\link[circlize]{posTransform.default}} for explanation
}
    \item{\code{col}}{Color of points. If there is only one numeric column, the length of \code{col} can be either one or number of rows of \code{region}.
 If there are more than one numeric column, the length of \code{col} can be either one or number of numeric columns.
 Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{pch}}{Type of points. Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{cex}}{Size of points. Settings are similar as \code{col}. Pass to \code{\link[circlize]{circos.points}}.
}
    \item{\code{bg}}{Background colors for points.
}
  }}
}
\value{
Object \linkS4class{ccGenomicCellGeom}
}
\description{
Object \linkS4class{ccGenomicCellGeom} will call the function \link{circos.genomicPoints} while drawing.
}
\examples{
data = generateRandomBed(nr =30, nc = 2)
all_chr = c("chr1","chr2","chr3","chr4","chr5","chr6","chr7","chr8","chr9","chr10","chr11","chr12","chr13","chr14","chr15","chr16","chr17","chr18","chr19","chr20","chr21","chr22","chrX","chrY")
cc = ccPlot(initMode = "initializeWithIdeogram", plotType=NULL)
t1 = ccGenomicTrack(data=data, numeric.column = 4)
cells1 = ccCells(sector.indexes = all_chr) + ccGenomicLines(numeric.column=2) + ccGenomicPoints(region=\(region,value){region}, value=\(region,value){value}, numeric.column=2)
t1 = t1 + cells1
show(cc+t1)
}
