% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{ccGenomicRainfall}
\alias{ccGenomicRainfall}
\title{Object generator for S4 class ccGenomicTrack}
\usage{
ccGenomicRainfall(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.genomicRainfall]{circlize::circos.genomicRainfall}}
  \describe{
    \item{\code{data}}{A bed-file-like data frame or a list of data frames.
}
    \item{\code{mode}}{How to calculate the distance of two neighbouring regions, pass to \code{\link[circlize]{rainfallTransform}}.
}
    \item{\code{ylim}}{ylim for rainfall plot track. If \code{normalize_to_width} is \code{FALSE}, the value should correspond to \code{log10(dist+1)},
 and if \code{normalize_to_width} is \code{TRUE}, the value should correspond to \code{log2(rel_dist)}.
}
    \item{\code{col}}{Color of points. It should be length of one. If \code{data} is a list, the length of \code{col}
 can also be the length of the list.
}
    \item{\code{pch}}{Style of points.
}
    \item{\code{cex}}{Size of points.
}
    \item{\code{normalize_to_width}}{If it is \code{TRUE}, the value is the relative distance divided by the width of the region.
}
  }}
}
\value{
Object \linkS4class{ccGenomicTrack}
}
\description{
Object \linkS4class{ccGenomicTrack} will call the function \link{circos.genomicRainfall} while drawing.
}
\examples{
NULL
}
