% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ccHeatmap}
\alias{ccHeatmap}
\title{Object generator for S4 class ccHeatmap}
\usage{
ccHeatmap(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.heatmap]{circlize::circos.heatmap}}
  \describe{
    \item{\code{mat}}{A matrix or a vector. The vector is transformed as a one-column matrix.}
    \item{\code{split}}{A categorical variable. It splits the matrix into a list of matrices.}
    \item{\code{col}}{If the values in the matrices are continuous, the color should be a color mapping generated by  \code{\link[circlize]{colorRamp2}}. If the values are characters, the color should be a named color vector.}
    \item{\code{na.col}}{Color for \code{NA} values.}
    \item{\code{cell.border}}{Border color of cells. A single scalar.}
    \item{\code{cell.lty}}{Line type of cell borders. A single scalar.}
    \item{\code{cell.lwd}}{Line width of cell borders. A single scalar.}
    \item{\code{bg.border}}{Color for background border.}
    \item{\code{bg.lty}}{Line type of the background border.}
    \item{\code{bg.lwd}}{Line width of the background border.}
    \item{\code{ignore.white}}{Whether to draw the white color?}
    \item{\code{cluster}}{whether to apply clustering on rows. The value can also be a \code{dendrogram}/\code{hclust} object or other objects that can be converted to with \code{\link[stats:dendrogram]{as.dendrogram}}.}
    \item{\code{clustering.method}}{Clustering method, pass to \code{\link[stats]{hclust}}.}
    \item{\code{distance.method}}{Distance method, pass to \code{\link[stats]{dist}}.}
    \item{\code{dend.callback}}{A callback function that is applied to the dendrogram in every sector.}
    \item{\code{dend.side}}{Side of the dendrograms relative to the heatmap track.}
    \item{\code{dend.track.height}}{Track height of the dendrograms.}
    \item{\code{rownames.side}}{Side of the row names relative to the heatmap track.}
    \item{\code{rownames.cex}}{Cex of row names.}
    \item{\code{rownames.font}}{Font of row names.}
    \item{\code{rownames.col}}{Color of row names.}
    \item{\code{show.sector.labels}}{Whether to show sector labels.}
    \item{\code{cell_width}}{Relative widths of heatmap cells.}
  }}
}
\value{
Object \linkS4class{ccHeatmap}
}
\description{
Object \linkS4class{ccHeatmap} will call the function \link[circlize:circos.heatmap]{circlize::circos.heatmap} while drawing.
}
\examples{
library(circlizePlus)
set.seed(123)
mat1 = rbind(cbind(matrix(rnorm(50*5, mean = 1), nr = 50),
                   matrix(rnorm(50*5, mean = -1), nr = 50)),
             cbind(matrix(rnorm(50*5, mean = -1), nr = 50),
                   matrix(rnorm(50*5, mean = 1), nr = 50))
            )
rownames(mat1) = paste0("R", 1:100)
colnames(mat1) = paste0("C", 1:10)
mat1 = mat1[sample(100, 100), ] # randomly permute rows
split = sample(letters[1:5], 100, replace = TRUE)
spilt = factor(split, levels = letters[1:5])
col_fun1 = colorRamp2(c(-2, 0, 2), c("blue", "white", "red"))
ccHeatmap(mat=mat1, split = split, col = col_fun1)
}
