% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{ccHeatmapLink}
\alias{ccHeatmapLink}
\title{Object generator for S4 class ccHeatmapLink}
\usage{
ccHeatmapLink(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.heatmap.link]{circlize::circos.heatmap.link}}
  \describe{
    \item{\code{row_from}}{The row index where the link starts. The value should be length 1. If you want to draw multiple links, put the function in a \code{for} loop.}
    \item{\code{row_to}}{The row index where the link ends.}
  }}
}
\value{
Object \linkS4class{ccHeatmapLink}
}
\description{
Object \linkS4class{ccHeatmapLink} will call the function \link{circos.heatmap.link} while drawing.
}
\examples{
library(circlizePlus)
set.seed(123)
mat = matrix(rnorm(100*10), nrow = 100)
rownames(mat) = paste0("R", 1:100)
col_fun = colorRamp2(c(-2, 0, 2), c("blue", "white", "red"))
cc = ccHeatmap(mat, col = col_fun, rownames.side = "outside")
link1 = ccHeatmapLink(10, 60)
cc + link1
}
