% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccLines}
\alias{ccLines}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccLines(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.lines]{circlize::circos.lines}}
  \describe{
    \item{\code{x}}{Data points on x-axis, measured in "current" data coordinate.}
    \item{\code{y}}{Data points on y-axis, measured in "current" data coordinate.}
    \item{\code{sector.index}}{Index for the sector.}
    \item{\code{track.index}}{Index for the track.}
    \item{\code{col}}{Line color.}
    \item{\code{lwd}}{Line width.}
    \item{\code{lty}}{Line style.}
    \item{\code{type}}{Line type, similar as \code{type} argument in \code{\link[graphics]{lines}}, but only in \code{c("l", "o", "h", "s")}}
    \item{\code{straight}}{Whether draw straight lines between points.}
    \item{\code{area}}{Whether to fill the area below the lines. If it is set to \code{TRUE}, \code{col} controls the filled color in the area and \code{border} controls color of the line.}
    \item{\code{area.baseline}}{deprecated, use \code{baseline} instead.}
    \item{\code{baseline}}{The base line to draw areas. By default it is the minimal of y-range (bottom). It can be a string or a number. If a string, it should be one of \code{bottom} and \code{top}. This argument also works if \code{type} is set to \code{h}.}
    \item{\code{border}}{color for border of the area.}
    \item{\code{pt.col}}{If \code{type} is "o", point color.}
    \item{\code{cex}}{If \code{type} is "o", point size.}
    \item{\code{pch}}{If \code{type} is "o", point type.}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.lines} while drawing.
}
\examples{
NULL
}
