% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\docType{class}
\name{ccLink-class}
\alias{ccLink-class}
\title{S4 class ccLink}
\description{
S4 class ccLink
}
\section{Slots}{

\describe{
\item{\code{func}}{character. Normally it is "circos.link".}

\item{\code{params}}{list. A \strong{named} list that stores the parameters of the function \link[circlize:circos.link]{circlize::circos.link} called by the backend.}
}}

\examples{
NULL
}
