% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccPolygon}
\alias{ccPolygon}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccPolygon(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.polygon]{circlize::circos.polygon}}
  \describe{
    \item{\code{x}}{Data points on x-axis}
    \item{\code{y}}{Data points on y-axis}
    \item{\code{sector.index}}{Index for the sector}
    \item{\code{track.index}}{Index for the track}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.polygon} while drawing.
}
\examples{
NULL
}
