% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccText}
\alias{ccText}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccText(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.text]{circlize::circos.text}}
  \describe{
    \item{\code{x}}{Data points on x-axis}
    \item{\code{y}}{Data points on y-axis}
    \item{\code{labels}}{Labels for each points}
    \item{\code{sector.index}}{Index for the sector}
    \item{\code{track.index}}{Index for the track}
    \item{\code{direction}}{deprecated, use \code{facing} instead.}
    \item{\code{facing}}{Facing of text. Please refer to vignette for different settings}
    \item{\code{niceFacing}}{Should the facing of text be adjusted to fit human eyes?}
    \item{\code{adj}}{offset for text. By default the text position adjustment is either horizontal or vertical in the canvas coordinate system. The "circular horizontal" offset can be set as a value in degree unit and the value should be wrapped by \code{\link[circlize]{degree}}.}
    \item{\code{cex}}{Font size}
    \item{\code{col}}{Font color}
    \item{\code{font}}{Font style}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.text} while drawing.
}
\examples{
NULL
}
