% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\docType{class}
\name{ccTrackGeom-class}
\alias{ccTrackGeom-class}
\title{S4 class ccTrackGeom}
\description{
Objectified representation of the R package circlize's plotting functions and corresponding parameters at the track level.
}
\section{Slots}{

\describe{
\item{\code{func}}{character. The name of the plot function in the R package circlize.}

\item{\code{params}}{list. A \strong{named} list that stores the parameters of the function called by the backend.}
}}

\examples{
NULL
}
