% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\name{ccTrackPoints}
\alias{ccTrackPoints}
\title{Add points on all sections of a single track.}
\usage{
ccTrackPoints(
  sectors,
  x,
  y,
  pch = par("pch"),
  col = par("col"),
  cex = par("cex"),
  bg = par("bg"),
  factors = sectors
)
}
\arguments{
\item{sectors}{A \code{\link{factor}} or a character vector which represents the categories of data}

\item{x}{Data points on x-axis}

\item{y}{Data points on y-axis}

\item{pch}{Point type}

\item{col}{Point color}

\item{cex}{Point size}

\item{bg}{backgrond color}

\item{factors}{The same as \code{sectors}. It will be removed in future versions. }
}
\value{
Object \linkS4class{ccTrackGeom}
}
\description{
Object \linkS4class{ccTrackGeom} will call the function \link[circlize:circos.trackPoints]{circlize::circos.trackPoints} while drawing.
}
\examples{
NULL
}
