% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-geom.R
\name{ccViolin}
\alias{ccViolin}
\title{Object generator for S4 class ccCellGeom}
\usage{
ccViolin(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[circlize:circos.violin]{circlize::circos.violin}}
  \describe{
    \item{\code{value}}{A numeric vector, a matrix or a list. If it is a matrix, boxplots are made by columns.}
    \item{\code{pos}}{Positions of the boxes.}
    \item{\code{violin_width}}{Width of violins.}
    \item{\code{col}}{Filled color of boxes.}
    \item{\code{border}}{Color for the border as well as the quantile lines.}
    \item{\code{lwd}}{Line width.}
    \item{\code{lty}}{Line style}
    \item{\code{show_quantile}}{Whether to show the quantile lines.}
    \item{\code{cex}}{Point size.}
    \item{\code{pch}}{Point type.}
    \item{\code{pt.col}}{Point color}
    \item{\code{max_density}}{The maximal density value across several violins. It is used to compare between violins.}
    \item{\code{sector.index}}{Index of sector.}
    \item{\code{track.index}}{Index of track.}
  }}
}
\value{
Object \linkS4class{ccCellGeom}
}
\description{
Object \linkS4class{ccCellGeom} will call the function \link{circos.violin} while drawing.
}
\examples{
NULL
}
