% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlizePlus-package.R
\docType{package}
\name{circlizePlus-package}
\alias{circlizePlus}
\alias{circlizePlus-package}
\title{circlizePlus: Using 'ggplot2' Feature to Write Readable R Code for Circular Visualization}
\description{
A wrapper for 'circlize'. All components are based on classes and objects. Users can use the addition symbol (+) to combine components for a circular visualization with 'ggplot2' style.The package is described in Zhang Z, Cao T, Huang Y and Xia Y (2025) \doi{10.3389/fgene.2025.1535368}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TianzeLab/circlizePlus}
  \item \url{https://tianzelab.github.io/circlizePlus/}
  \item \doi{10.3389/fgene.2025.1535368}
  \item Report bugs at \url{https://github.com/TianzeLab/circlizePlus/issues}
}

}
\author{
\strong{Maintainer}: Tianze Cao \email{hnrcao@qq.com} (\href{https://orcid.org/0000-0003-4308-1878}{ORCID})

Authors:
\itemize{
  \item Zheyu Zhang \email{2054906893@qq.com}
}

}
\keyword{internal}
