% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data-set}
\alias{data-set}
\title{Data set in circlizePlus}
\source{
\url{https://www.frontiersin.org/journals/genetics/articles/10.3389/fgene.2025.1535368/full#h5}
}
\description{
Example can be found in "4.2 Example 2: comparison of two pieces of code that use circlizePlus and circlize to implement the same requirements, respectively"
\if{html}{\out{<br>}}allele_frequency: A dataframe with 25,000 rows and 4 columns
\if{html}{\out{<br>}}copy_number: A dataframe with 24,788 rows and 5 columns
\if{html}{\out{<br>}}junctions: A dataframe with 2414 rows and 28 columns
\if{html}{\out{<br>}}r1: A dataframe with 2414 rows and 3 columns
\if{html}{\out{<br>}}r2: A dataframe with 2414 rows and 3 columns
}
\examples{
data("example2")

}
